// src/03_canisters/yumi/yumi_core/module_3242c807/core_3242c807.did.d.ts
export type Img = string;
// ...
export type OfferId__2 = bigint;
// ...
export type ProfileErr = { alreadyCreate: null } | { noProfile: null } | { defaultAccount: null };
export interface ProfileLet {
    bio: string;
    userName: string;
    created: Array<TokenIdentifier__4>;
    favorited: Array<TokenIdentifier__4>;
    userId: UserId__2;
    time: Time;
    banner: Img;
    notification: Array<string>;
    offersReceived: Array<OfferId__2>;
    collections: Array<Principal>;
    email: string;
    collected: Array<TokenIdentifier__4>;
    offersMade: Array<OfferId__2>;
    followeds: Array<UserId__2>;
    followers: Array<UserId__2>;
    avatar: Img;
}
// ...
export type Result_2 = { ok: ProfileLet } | { err: ProfileErr };
// ...
export type Time = bigint;
export type TokenIdentifier = string;
export type TokenIdentifier__1 = string;
export type TokenIdentifier__2 = string;
export type TokenIdentifier__3 = string;
export type TokenIdentifier__4 = string;
// ...
export type UserId__2 = Principal;